MDL-78597 core_course: fix content item tests covering get_all cases
Fixes two tests which use the lti module as an example of a core mod
implementing the 'get_all_content_items' callback. The setup of the mod
in these tests, including course and site tools, needed to be updated
in light of the removal of the 'external tool' content item.