MDL-67271 componentlibrary: Update fontawesomeicons
Regenerate the 'fontawesomeicons.json' file to enforce SVG rendering for
the icons.
The addition of '$CFG->svgicons = true;' in the config.php file addressed this.
It's important to note that '$CFG->svgicons' will eventually be deprecated as part
of MDL-78993. Consequently, I refrained from updating 'readme_moodle.txt,' as SVG
icons will become the default value once the removal is executed.