MDL-79248 mod_quiz: Behat scenario interacts incorrectly with select
This test was using the click operator to select an option value, rather
than selecting the value with form field steps.

In addition I found that sometimes the way in which a WS is used to
make the change before a refresh leads to a random failure if the page
navigation happens too slowly.