MDL-79638 phpunit: Remove mysql/mariadb hack
This hack was introduced to work around a bug in MySQL 5.6.14 and
MariaDB at the time.

https://bugs.mysql.com/bug.php?id=69882

It was addressed a few months later in 5.6.16, and 5.7.4.
MariaDB merged version 5.6.16 of MySQL's InnoDB engine in MariaDB
10.0.11 and got hte patch from there.

Moodle has required MariaDB 10.2.29 since Moodle 3.9 and it is therefore
safe to remove these hacks for MariaDB.
MySQL is also affected for version 5.7.0 >= MySQL < 5.7.4.