MDL-77166 core: Auto close pull requests @ GitHub
Note that we have also enabled it to be executed
via manual dispatch to be able to close currently
existing PRs.

Once everything existing is closed, it will take on
any new PR arriving. Then, we can remove the manual
dispatching, although it's not critical to leave it
there in case we want to use that workflow with other
repos.