MDL-68806 question: Set first attempt step timecreated
Previously, the timecreated for the first step of any question attempt
was set to the current time when the quiz attempt was created. With quiz
attempts being pre-created, this means the question attempts looking
like they were started before the quiz attempt. It also limits the
ability to analyse when a student actually started each question.

This change adds a new class constant,
question_attempt_step::TIMECREATED_ON_FIRST_RENDER. Passing this to the
question_attempt_step constructor will cause the timecreated to be set
to the time when the question attempt is first rendered, rather than
when the step is created.