NOBUG: Avoid already closed PRs to get any comment
This is an adjustment for MDL-77166, where the auto-close PRs
workflow was introduced.

That workflow does close and lock any PR in the moodle/moodle.git repo.

We have detected that, when proceeding to lock the PRs it's
always adding a comment. But we only want that when closing the
PRs (as they are open). When the only task is to lock them (normally
old PRs that are already closed by not locked) we don't want any
comment to be added (so no notification/spam happens on those old PRs).

Luckily there is a setting for it: skip-closed-pr-comment: true

So applying for it. Note that we have run various tests in fork
repository and it seems to be working ok. Only open PRs got the
comment and closed PRs were silently locked.